Backend/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
Pillow==11.0.0
opencv-python==4.10.0.84
pytesseract==0.3.13